Refactor UQD and Device Management in tagDisp
- Enhanced UQD class with buffer clearing and improved process management.
- Updated UQDSettingsGroup to include start/stop functionality for the timetagger.
- Simplified device selection and remote connection handling in tagdisp.py.
- Removed deprecated tagdisp_device.py and integrated its functionality into DeviceBox.
- Improved UI responsiveness and added clear buffer functionality in the settings.
- Streamlined data handling in SimpleDisplay and updated method signatures for clarity.
